did a lot of work

+++ 300/302.cpp
@@ -1,9 +1,8 @@
 #include <iostream>
-#include <cmath>
 using namespace std;
 
 int main() {
-    float x, y, z;
+    double x, y, z;
     cin >> x >> y >> z;
 
     if (x <= y && y <= z) {
